Lo and behold, beautiful configurations and tests to back it up
