+++ yucca/training/managers/YuccaManager.py
@@ -3,8 +3,10 @@ import torch
 from typing import Literal, Union, Optional
 from yucca.training.augmentation.YuccaAugmentationComposer import YuccaAugmentationComposer
 from yucca.training.configuration.split_data import get_split_config
+from yucca.training.configuration.configure_task import get_task_config
 from yucca.training.configuration.configure_callbacks import get_callback_config
-from yucca.training.configuration.configure_paths_and_plans import get_path_and_plan_config
+from yucca.training.configuration.configure_paths_and_version import get_path_and_version_config
+from yucca.training.configuration.configure_plans import get_plan_config
 from yucca.training.data_loading.YuccaDataModule import YuccaDataModule
 from yucca.training.lightning_modules.YuccaLightningModule import YuccaLightningModule
 from yucca.training.configuration.input_dimensions import get_input_dims
@@ -83,7 +85,6 @@ class YuccaManager:
 
         # default settings
         self.max_vram = 12
-        self.is_initialized = False
 
         # Trainer settings
         self.train_batches_per_step = 250
@@ -100,45 +101,66 @@ class YuccaManager:
     ):
         # Here we configure the outpath we will use to store model files and metadata
         # along with the path to plans file which will also be loaded.
-        self.path_and_plan_config = get_path_and_plan_config(
-            ckpt_path=self.ckpt_path,
+        task_config = get_task_config(
             continue_from_most_recent=self.continue_from_most_recent,
             manager_name=self.name,
             model_dimensions=self.model_dimensions,
             model_name=self.model_name,
-            planner=self.planner,
+            planner_name=self.planner,
             split_idx=self.split_idx,
             task=self.task,
         )
 
-        splits = get_split_config(self.path_and_plan_config.train_data_dir, self.path_and_plan_config.task)
+        self.path_config = get_path_and_version_config(
+            continue_from_most_recent=task_config.continue_from_most_recent,
+            manager_name=task_config.manager_name,
+            model_dimensions=task_config.model_dimensions,
+            model_name=task_config.model_name,
+            planner_name=task_config.planner_name,
+            split_idx=task_config.split_idx,
+            task=task_config.task,
+        )
+
+        plan_config = get_plan_config(
+            ckpt_path=self.ckpt_path,
+            continue_from_most_recent=task_config.continue_from_most_recent,
+            plans_path=self.path_config.plans_path,
+            version=self.path_config.version,
+            version_dir=self.path_config.version_dir,
+        )
+
+        splits = get_split_config(train_data_dir=self.path_config.train_data_dir, task=task_config.task)
         input_dims = get_input_dims(
-            self.path_and_plan_config.plans,
-            self.model_dimensions,
-            self.path_and_plan_config.num_classes,
-            self.path_and_plan_config.model_name,
-            self.max_vram,
-            self.patch_size,
+            plan=plan_config.plans,
+            model_dimensions=task_config.model_dimensions,
+            num_classes=plan_config.num_classes,
+            model_name=task_config.model_name,
+            max_vram=self.max_vram,
+            patch_size=self.patch_size,
         )
 
         augmenter = YuccaAugmentationComposer(
             patch_size=input_dims.patch_size,
             is_2D=True if self.model_dimensions == "2D" else False,
-            use_preset_for_task_type=self.path_and_plan_config.task_type,
+            use_preset_for_task_type=plan_config.task_type,
         )
 
         callback_config = get_callback_config(
-            task=self.task,
-            save_dir=self.path_and_plan_config.save_dir,
-            version_dir=self.path_and_plan_config.version_dir,
-            version=self.path_and_plan_config.version,
+            task=task_config.task,
+            save_dir=self.path_config.save_dir,
+            version_dir=self.path_config.version_dir,
+            version=self.path_config.version,
             disable_logging=self.disable_logging,
             prediction_output_dir=prediction_output_dir,
             profile=self.profile,
             save_softmax=save_softmax,
         )
         self.model_module = YuccaLightningModule(
-            config=self.path_and_plan_config.lm_hparams() | splits.lm_hparams() | input_dims.lm_hparams(),
+            config=task_config.lm_hparams()
+            | self.path_config.lm_hparams()
+            | plan_config.lm_hparams()
+            | splits.lm_hparams()
+            | input_dims.lm_hparams(),
             loss_fn=self.loss,
             stage=stage,
             step_logging=self.step_logging,
@@ -146,20 +168,21 @@ class YuccaManager:
         )
 
         self.data_module = YuccaDataModule(
-            splits=splits,
-            split_idx=self.split_idx,
-            input_dims=input_dims,
             composed_train_transforms=augmenter.train_transforms,
             composed_val_transforms=augmenter.val_transforms,
-            config=self.path_and_plan_config,
+            input_dims=input_dims,
             num_workers=self.num_workers,
+            plan_config=plan_config,
             pred_data_dir=pred_data_dir,
             pre_aug_patch_size=augmenter.pre_aug_patch_size,
+            splits=splits,
+            split_idx=task_config.split_idx,
+            train_data_dir=self.path_config.train_data_dir,
         )
 
         self.trainer = L.Trainer(
             callbacks=callback_config.callbacks,
-            default_root_dir=self.path_and_plan_config.save_dir,
+            default_root_dir=self.path_config.save_dir,
             limit_train_batches=self.train_batches_per_step,
             limit_val_batches=self.val_batches_per_step,
             logger=callback_config.loggers,
@@ -181,7 +204,7 @@ class YuccaManager:
     def run_finetuning(self):
         self.initialize(stage="fit")
         self.model_module.load_state_dict(
-            torch.load(self.path_and_plan_config.ckpt_path, map_location=torch.device("cpu"))["state_dict"], strict=False
+            torch.load(self.path_config.ckpt_path, map_location=torch.device("cpu"))["state_dict"], strict=False
         )
         self.trainer.fit(
             model=self.model_module,
